+++ android/build.gradle
@@ -51,7 +51,7 @@ android {
     }
 
     dependencies {
-        implementation("com.coralogix:android-sdk:2.1.0")
+        implementation("com.coralogix:android-sdk:2.1.1")
         coreLibraryDesugaring("com.android.tools:desugar_jdk_libs:2.1.3")
         testImplementation("org.jetbrains.kotlin:kotlin-test")
         testImplementation("org.mockito:mockito-core:5.12.0")
