+++ android/src/main/kotlin/com/coralogix/flutter/plugin/extensions/ListExtensions.kt
@@ -1,0 +1,5 @@
+package com.coralogix.flutter.plugin.extensions
+
+internal fun List<*>.toStringList(): List<String> {
+    return mapNotNull { it as? String }
+}
