+++ android/src/main/kotlin/com/coralogix/flutter/plugin/extensions/ResultExtensions.kt
@@ -1,0 +1,13 @@
+package com.coralogix.flutter.plugin.extensions
+
+import io.flutter.plugin.common.MethodChannel.Result
+
+internal fun Result.invalidArgumentsError() {
+    error("4", "Arguments is null or empty", null)
+}
+
+internal fun Result.error(description: String) {
+    error("4", description, null)
+}
+
+internal fun Result.success() = success("")
