+++ android/src/main/kotlin/com/coralogix/flutter/plugin/mappers/IMapper.kt
@@ -1,0 +1,5 @@
+package com.coralogix.flutter.plugin.mappers
+
+interface IMapper<in I, out O> {
+    fun map(input: I): O
+}
